Ycz/glip inf (#13)
* add vlfuse file

* mv bert fuse classes to vlfuse helper
